moving tests to another file
